# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API - Cadastro de Usuários",
    "description": "Documentação das rotas de cadastro e consulta de instituições beneficiárias e pessoas físicas.",
    "version": "1.0.0"
  },
  "paths": {
    "/registerIB": {
      "post": {
        "summary": "Cadastro de Instituição Beneficiária",
        "description": "Cadastra uma nova instituição beneficiária.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nameInc": { "type": "string" },
                  "emailInc": { "type": "string", "format": "email" },
                  "passwordInc": { "type": "string", "format": "password" },
                  "cnpjInc": { "type": "string" },
                  "locationInc": { "type": "string" },
                  "telInc": { "type": "string" }
                },
                "required": ["nameInc", "emailInc", "passwordInc", "cnpjInc", "locationInc", "telInc"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "✅ Cadastro realizado! Verifique seu e-mail." },
          "400": { "description": "Preencha todos os campos!" },
          "409": { "description": "❌ E-mail ou CNPJ já cadastrado." },
          "500": { "description": "Erro ao cadastrar instituição." }
        }
      }
    },
...
    "/instituicao": {
      "get": {
        "summary": "Consulta de Instituição",
        "description": "Busca uma instituição beneficiária por ID, e-mail ou CNPJ.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "email",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "cnpj",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Consulta realizada com sucesso." },
          "400": { "description": "Parâmetro inválido." },
          "404": { "description": "Instituição não encontrada." },
          "500": { "description": "Erro no servidor." }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
      "/verificar-emailIB": {
      "get": {
        "summary": "Verificação de E-mail da Instituição Beneficiária",
        "description": "Verifica o e-mail de uma instituição beneficiária por meio de um token JWT enviado via query string.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
            },
            "description": "Token JWT enviado no e-mail para verificação da instituição beneficiária."
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail verificado com sucesso.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "E-mail verificado com sucesso!"
                }
              }
            }
          },
          "400": {
            "description": "Token ausente.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Token ausente."
                }
              }
            }
          },
          "401": {
            "description": "Token inválido ou expirado.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Token inválido ou expirado."
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao tentar verificar o e-mail.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Erro interno."
                }
              }
            }
          }
        }
      }
    },
...
    "/cadastrar-itemI": {
      "post": {
        "summary": "Cadastrar Pedido por Instituição Beneficiária",
        "description": "Registra um novo pedido no sistema feito por uma instituição beneficiária. Todos os campos são obrigatórios.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name_item",
                  "description",
                  "quantity_item",
                  "category",
                  "urgencia_enum",
                  "locate"
                ],
                "properties": {
                  "id": {
                    "type": "integer",
                    "example": 12,
                    "description": "ID da instituição beneficiária"
                  },
                  "name_item": {
                    "type": "string",
                    "example": "Roupa de frio",
                    "description": "Nome do item solicitado"
                  },
                  "description": {
                    "type": "string",
                    "example": "Roupas de inverno para população em situação de rua",
                    "description": "Descrição detalhada do pedido"
                  },
                  "quantity_item": {
                    "type": "integer",
                    "example": 50,
                    "description": "Quantidade solicitada"
                  },
                  "category": {
                    "type": "string",
                    "example": "Vestuário",
                    "description": "Categoria do item"
                  },
                  "urgencia_enum": {
                    "type": "string",
                    "enum": ["Urgente", "Alta", "Média", "Baixa"],
                    "example": "Alta",
                    "description": "Nível de urgência do pedido"
                  },
                  "locate": {
                    "type": "string",
                    "example": "Av. Central, 456 - Bairro Centro",
                    "description": "Endereço de entrega"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido cadastrado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Pedido realizado com sucesso."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição mal formatada ou dados inválidos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID inválido!"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao tentar cadastrar o pedido.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno ao criar o pedido."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/pedidosI": {
      "get": {
        "summary": "Listar pedidos por instituição",
        "description": "Retorna todos os pedidos feitos por uma instituição com base no ID.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "ID da instituição",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "instituicao_id": { "type": "integer" },
                      "name_item": { "type": "string" },
                      "description": { "type": "string" },
                      "quantity_item": { "type": "integer" },
                      "category": { "type": "string" },
                      "urgencia_enum": { "type": "string" },
                      "locate": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro no servidor"
          }
        }
      }
    },
...
"/deleteinstituicao": {
  "delete": {
    "summary": "Deletar instituição beneficiária",
    "description": "Remove uma instituição beneficiária e seus pedidos associados do banco de dados. Requer autenticação com token JWT.",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "parameters": [
      {
        "name": "id",
        "in": "query",
        "required": true,
        "description": "ID da instituição beneficiária a ser deletada",
        "schema": {
          "type": "string"
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Instituição deletada com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Instituição deletada com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "ID da instituição não fornecido"
      },
      "404": {
        "description": "Instituição não encontrada"
      },
      "500": {
        "description": "Erro ao excluir pedidos ou instituição"
      }
    }
  }
},
...
"/beneficiarioup/{id}": {
  "put": {
    "summary": "Atualizar dados da instituição beneficiária",
    "description": "Atualiza nome, email, CNPJ, localização e telefone da instituição beneficiária, exigindo confirmação da senha atual.",
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID da instituição beneficiária a ser atualizada",
        "schema": {
          "type": "string"
        }
      }
    ],
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "required": ["nameInc", "emailInc", "passwordInc", "cnpjInc", "locationInc", "telInc"],
            "properties": {
              "nameInc": {
                "type": "string",
                "description": "Nome da instituição"
              },
              "emailInc": {
                "type": "string",
                "format": "email",
                "description": "Email da instituição"
              },
              "passwordInc": {
                "type": "string",
                "description": "Senha atual para validação"
              },
              "cnpjInc": {
                "type": "string",
                "description": "CNPJ da instituição"
              },
              "locationInc": {
                "type": "string",
                "description": "Localização da instituição"
              },
              "telInc": {
                "type": "string",
                "description": "Telefone da instituição"
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Instituição atualizada com sucesso",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "✅ Instituição atualizada com sucesso"
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Campos obrigatórios não preenchidos"
      },
      "401": {
        "description": "Senha atual incorreta"
      },
      "404": {
        "description": "Instituição não encontrada"
      },
      "500": {
        "description": "Erro no servidor"
      }
    }
  }
},
...
}
}